requirements.txt
numpy>=1.20.0,<2.0.0
pandas>=1.3.0
matplotlib>=3.4.0
seaborn>=0.11.0
scikit-learn>=0.24.0
tensorflow==2.16.1
gymnasium>=0.28.0
ta>=0.7.0
python-binance>=1.0.0
python-dotenv>=0.19.0
tqdm>=4.60.0
streamlit>=1.0.0
plotly>=5.0.0
